fix: circular dependancy issues
