+++ harness/helpers/helpers.ts
@@ -1,5 +1,5 @@
 import { Interceptor, Scope } from 'nock'
-import { Capabilities, SDKCapabilities, Sdks } from '../types'
+import { SDKCapabilities } from '../types'
 import { getServerScope, resetServerScope } from '../nock'
 import { v4 as uuidv4 } from 'uuid'
 import {
@@ -49,51 +49,6 @@ export const getConnectionStringForProxy = (proxy: string) => {
 
 let currentClient: BaseTestClient
 const clientTestNameMap: Record<string, string> = {}
-
-// Extract the list of SDKs to test from the environment variable SDKS_TO_TEST
-// It can be formatted as a JSON array or a comma-separated list.
-// The SDKs are returned as their human-readable names from the Sdks enum, not the enum keys.
-export const getSDKs = () => {
-    const SDKS_TO_TEST = process.env.SDKS_TO_TEST
-
-    try {
-        return JSON.parse(SDKS_TO_TEST ?? '').map((sdk) => Sdks[sdk])
-    } catch (e) {
-        if (SDKS_TO_TEST && Sdks[SDKS_TO_TEST]) {
-            return [Sdks[SDKS_TO_TEST]]
-        } else if (SDKS_TO_TEST) {
-            return SDKS_TO_TEST.split(',')
-                .map((sdk) => Sdks[sdk])
-                .filter((sdkName) => sdkName !== undefined)
-        } else {
-            console.warn('No specified SDKs to test, running all tests')
-            return Object.values(Sdks)
-        }
-    }
-}
-
-export const getCapabilities = () => {
-    const SDK_CAPABILITIES = process.env.SDK_CAPABILITIES
-
-    try {
-        return JSON.parse(SDK_CAPABILITIES ?? '').map(
-            (sdk) => Capabilities[sdk],
-        )
-    } catch (e) {
-        if (SDK_CAPABILITIES && Capabilities[SDK_CAPABILITIES]) {
-            return [Capabilities[SDK_CAPABILITIES]]
-        } else if (SDK_CAPABILITIES) {
-            return SDK_CAPABILITIES.split(',')
-                .map((capability) => Capabilities[capability])
-                .filter((capability) => capability !== undefined)
-        } else {
-            console.warn(
-                'No specified SDK Capabilities to test, running all tests',
-            )
-            return Object.values(Capabilities)
-        }
-    }
-}
 
 export const getSDKName = () => global.JEST_PROJECT_SDK_TO_TEST
