+++ harness/types/capabilities.test.ts
@@ -1,5 +1,3 @@
-import { Sdks } from './sdks'
-
 describe('SDKCapabilities Unit Tests', () => {
     afterEach(() => {
         jest.resetModules()
@@ -7,28 +5,32 @@ describe('SDKCapabilities Unit Tests', () => {
 
     it('should load capabilities from env variables', () => {
         process.env.SDK_CAPABILITIES = 'edgeDB,cloud,clientCustomData,sse'
-        process.env.SDKS_TO_TEST = 'NodeJS,Python'
-        const { SDKCapabilities } = jest.requireActual('../capabilities')
+        process.env.SDKS_TO_TEST = 'nodejs,python'
+        const { SDKCapabilities } = jest.requireActual('./capabilities')
         expect(SDKCapabilities).toEqual({
-            NodeJS: ['edgeDB', 'cloud', 'clientCustomData', 'sse'],
-            Python: ['edgeDB', 'cloud', 'clientCustomData', 'sse'],
+            NodeJS: ['EdgeDB', 'CloudBucketing', 'ClientCustomData', 'SSE'],
+            Python: ['EdgeDB', 'CloudBucketing', 'ClientCustomData', 'SSE'],
         })
     })
 
     it('should load capabilities from env variables as JSON', () => {
         process.env.SDK_CAPABILITIES =
             '["edgeDB","cloud","clientCustomData","sse"]'
-        process.env.SDKS_TO_TEST = '["NodeJS","Python"]'
-        const { SDKCapabilities } = jest.requireActual('../capabilities')
+        process.env.SDKS_TO_TEST = '["nodejs","python"]'
+        const { SDKCapabilities } = jest.requireActual('./capabilities')
         expect(SDKCapabilities).toEqual({
-            NodeJS: ['edgeDB', 'cloud', 'clientCustomData', 'sse'],
-            Python: ['edgeDB', 'cloud', 'clientCustomData', 'sse'],
+            NodeJS: ['EdgeDB', 'CloudBucketing', 'ClientCustomData', 'SSE'],
+            Python: ['EdgeDB', 'CloudBucketing', 'ClientCustomData', 'SSE'],
         })
     })
 
     it('should use default capabilities if env variables are not set', () => {
         process.env.SDK_CAPABILITIES = 'edgeDB,cloud,clientCustomData,sse'
-        const { SDKCapabilities } = jest.requireActual('../capabilities')
-        expect(Object.keys(SDKCapabilities)).toEqual(Object.values(Sdks))
+        delete process.env.SDKS_TO_TEST
+        const { SDKCapabilities } = jest.requireActual('./capabilities')
+        const { Sdks } = jest.requireActual('./sdks')
+        expect(Object.keys(SDKCapabilities)).toEqual(
+            expect.arrayContaining(Object.values(Sdks)),
+        )
     })
 })
