+++ harness/types/capabilities.ts
@@ -1,4 +1,4 @@
-import { getCapabilities, getSDKs } from '../helpers'
+import { getSDKs } from './sdks'
 
 export const Capabilities = {
     edgeDB: 'EdgeDB',
@@ -57,6 +57,29 @@ let sdkCapabilities: { [key: string]: string[] } = {
     ],
     Ruby: [Capabilities.clientCustomData],
     PHP: [Capabilities.cloudProxy],
+}
+
+export const getCapabilities = () => {
+    const SDK_CAPABILITIES = process.env.SDK_CAPABILITIES
+
+    try {
+        return JSON.parse(SDK_CAPABILITIES ?? '').map(
+            (sdk) => Capabilities[sdk],
+        )
+    } catch (e) {
+        if (SDK_CAPABILITIES && Capabilities[SDK_CAPABILITIES]) {
+            return [Capabilities[SDK_CAPABILITIES]]
+        } else if (SDK_CAPABILITIES) {
+            return SDK_CAPABILITIES.split(',')
+                .map((capability) => Capabilities[capability])
+                .filter((capability) => capability !== undefined)
+        } else {
+            console.warn(
+                'No specified SDK Capabilities to test, running all tests',
+            )
+            return Object.values(Capabilities)
+        }
+    }
 }
 
 if (process.env.SDK_CAPABILITIES && process.env.SDKS_TO_TEST) {
