+++ jest.config.ts
@@ -1,4 +1,4 @@
-import { getSDKs } from './harness/helpers/helpers'
+import { getSDKs } from './harness/types'
 
 const commonConfig = {
     preset: '@eresearchqut/jest-testcontainers',
